+++ p2panda-auth/src/traits/dgm.rs
@@ -1,112 +1,91 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
 use std::collections::HashSet;
-use std::error::Error;
+use std::fmt::Debug;
 
 use crate::Access;
 use crate::group::GroupMember;
-use crate::traits::{IdentityHandle, OperationId, Orderer};
+use crate::traits::{IdentityHandle, OperationId};
 
 /// Decentralised group membership (DGM) API for managing membership of a single group.
-pub trait Group<ID, OP, C, ORD>
+pub trait Groups<ID, OP, C, MSG>
 where
     ID: IdentityHandle,
     OP: OperationId,
-    // TODO: Do we strictly need the orderer here? Could it rather be a generic message?
-    // We might not actually need to know anything about the message type, only in the `Orderer`.
-    // In the _implementation_ we'd say it's an `ORD::Operation` but not here (move that knowledge
-    // into the implementation.
-    ORD: Orderer<ID, OP, Self::Action>,
 {
-    type State;
-    type Action;
-    type Error: Error;
+    type Error: Debug;
 
     /// Creates a new group, returning the updated state and the creation operation message.
     ///
     /// The group state must first be initialised by calling `init()` before this function is
     /// called.
     fn create(
-        &self,
+        &mut self,
         group_id: ID,
         initial_members: Vec<(GroupMember<ID>, Access<C>)>,
-    ) -> Result<(Self::State, ORD::Operation), Self::Error>;
-
-    /// Initialise the group by processing a remotely-authored `create` message.
-    ///
-    /// The group state must first be initialised by calling `init()` before this function is
-    /// called. The `group_id` is extracted from the operation itself.
-    fn create_from_remote(
-        &self,
-        remote_operation: ORD::Operation,
-    ) -> Result<Self::State, Self::Error>;
+    ) -> Result<MSG, Self::Error>;
 
     /// Process a remotely-authored group action message.
-    fn receive_from_remote(
-        y: Self::State,
-        remote_operation: ORD::Operation,
-    ) -> Result<Self::State, Self::Error>;
+    fn receive_from_remote(&mut self, remote_operation: MSG) -> Result<(), Self::Error>;
 
     /// Add a member to the group.
     ///
     /// The updated state is returned, as well as the `add` operation. The operation should be
     /// shared with remote peers so they can update their group state accordingly.
     fn add(
-        y: Self::State,
+        &mut self,
+        group_id: ID,
         adder: ID,
         added: ID,
         access: Access<C>,
-    ) -> Result<(Self::State, ORD::Operation), Self::Error>;
+    ) -> Result<MSG, Self::Error>;
 
     /// Removes a member from the group.
-    fn remove(
-        y: Self::State,
-        remover: ID,
-        removed: ID,
-    ) -> Result<(Self::State, ORD::Operation), Self::Error>;
+    fn remove(&mut self, group_id: ID, remover: ID, removed: ID) -> Result<MSG, Self::Error>;
 
     /// Promote a member to the given access level.
     fn promote(
-        y: Self::State,
+        &mut self,
+        group_id: ID,
         promoter: ID,
         promoted: ID,
         access: Access<C>,
-    ) -> Result<(Self::State, ORD::Operation), Self::Error>;
+    ) -> Result<MSG, Self::Error>;
 
     /// Demote a member to the given access level.
     fn demote(
-        y: Self::State,
+        &mut self,
+        group_id: ID,
         demoter: ID,
         demoted: ID,
         access: Access<C>,
-    ) -> Result<(Self::State, ORD::Operation), Self::Error>;
+    ) -> Result<MSG, Self::Error>;
 }
 
 /// Interface for querying group membership and access levels.
 pub trait GroupMembership<ID, OP, C> {
-    type State;
-    type Error: Error;
+    type Error: Debug;
 
     /// Query the current access level of the given member.
     ///
     /// The member is expected to be a "stateless" individual, not a "stateful" group.
-    fn access(y: &Self::State, member: &ID) -> Result<Access<C>, Self::Error>;
+    fn access(&self, group_id: ID, member: ID) -> Result<Access<C>, Self::Error>;
 
     /// Query group membership.
-    fn member_ids(y: &Self::State) -> Result<HashSet<ID>, Self::Error>;
+    fn member_ids(&self, group_id: ID) -> Result<HashSet<ID>, Self::Error>;
 
     /// Return `true` if the given ID is an active member of the group.
-    fn is_member(y: &Self::State, member: &ID) -> Result<bool, Self::Error>;
+    fn is_member(&self, group_id: ID, member: ID) -> Result<bool, Self::Error>;
 
     /// Return `true` if the given member is currently assigned the `Pull` access level.
-    fn is_puller(y: &Self::State, member: &ID) -> Result<bool, Self::Error>;
+    fn is_puller(&self, group_id: ID, member: ID) -> Result<bool, Self::Error>;
 
     /// Return `true` if the given member is currently assigned the `Read` access level.
-    fn is_reader(y: &Self::State, member: &ID) -> Result<bool, Self::Error>;
+    fn is_reader(&self, group_id: ID, member: ID) -> Result<bool, Self::Error>;
 
     /// Return `true` if the given member is currently assigned the `Write` access level.
-    fn is_writer(y: &Self::State, member: &ID) -> Result<bool, Self::Error>;
+    fn is_writer(&self, group_id: ID, member: ID) -> Result<bool, Self::Error>;
 
     /// Return `true` if the given member is currently assigned the `Manage` access level.
-    fn is_manager(y: &Self::State, member: &ID) -> Result<bool, Self::Error>;
+    fn is_manager(&self, group_id: ID, member: ID) -> Result<bool, Self::Error>;
 }